Destroy WebTorrent on app exit
In order to delete the files from the tmp folder, we have to make sure the WebTorrent is destroyed. This resolves problems with not being able to do unlink files.

We don't have to know the outcome of the callback, since when it's an error, the WebTorrent is already closed.